refactor(1013): extract APIDataFetcher to src/api/ (SC-001, position 21)
Cat B fresh-extraction: move APIDataFetcher class (~328 lines) from
MistHelper.py to new src/api/api_data_fetcher.py, keeping the class
name reachable via a re-export alias so the 12 in-monolith call sites
continue to resolve through MistHelper.APIDataFetcher.

- src/api/api_data_fetcher.py: extracted class with lazy
  importlib.import_module("MistHelper") for access to apisession,
  ConfigUtils, RateLimitingUtils, DataExporter, DataProcessingUtils,
  and retry-constant globals.
- MistHelper.py: comment-only breadcrumb at former class site;
  alphabetical re-export block between src.analytics and src.audit.
- pyproject.toml: add src.api.* to Wave-2 mypy relaxation block;
  add src/api/api_data_fetcher.py to coverage.omit (facade dominated
  by live API + interactive-input paths).

Local gate: black + ruff clean; py_compile OK; MistHelper.py --test
= 66 passed / 0 failed / 131 skipped (matches baseline).
